apache2: work around bug in apr-1-config/apu-1-config scripts
  (original commit: dff2c1bd5291e95b35a102c1b5fdbe78aa627c41)

The apr-1-config and apu-1-config scripts (installed into the sysroot
by apr and apr-util respectively) report incorrect paths with current
OE-Core - see:

  http://bugzilla.yoctoproject.org/show_bug.cgi?id=3267

Since OE-Core is frozen at the moment, work around the issue by taking
a copy of the scripts and fixing them up. Additionally we need to not
mangle the libtool name as apr installs it into the sysroot as just
'libtool'.

This fixes building apache2 with rm_work enabled.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>