ajenti: remove
  (original commit: c9236a3665494fb53f8ab46cdf51007bebde7f3e)

In recent versions, upstream has decided to place additional
restrictions on commercial use beyond a standard open source license
(LGPLv3) [1]. This makes it hard to set a LICENSE value that is easily
understood. Of course, as the authors, they have the right to decide
what licensing terms they wish to distribute their project under, and we
could always set LICENSE_FLAGS to denote the extra terms, but this is
somewhat messy and personally I feel less inclined to continue
maintaining this recipe in meta-webserver now, especially since I
originally put it together on my own time. At the moment due to a
branch/commit mismatch it is no longer fetching in any case.

(If someone wants to resurrect this recipe in another layer, they are
more than welcome to do so.)

[1] http://support.ajenti.org/topic/351265-clarify-licensing/

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>